requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
matplotlib>=3.7.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
hydra-core>=1.3.0